+++ click_extra/sphinx/__init__.py
@@ -36,13 +36,13 @@ from sphinx.highlighting import PygmentsBridge
 from sphinx.util import logging
 
 from .. import __version__
-from ..pygments import AnsiHtmlFormatter
-from . import manpages, matrix
-from ._base import (
+from ..blocks import (
     marker_res as marker_res,
     replace_region as replace_region,
     update_blocks as update_blocks,
 )
+from ..pygments import AnsiHtmlFormatter
+from . import manpages, matrix
 from .alerts import convert_github_alerts
 from .click import ClickDomain, cleanup_runner
 from .python import (